introduce LocationProtocol
