+++ Sources/OpenWeatherKit/Public/WeatherService.swift
@@ -115,7 +115,7 @@ final public class WeatherService: Sendable {
     /// - Throws: Weather data error `WeatherError`
     /// - Returns: The aggregate weather.
     ///
-    final public func weather(for location: Location, countryCode: String) async throws -> Weather {
+    final public func weather(for location: LocationProtocol, countryCode: String) async throws -> Weather {
         let proxy = try await networkClient.fetchWeather(
             location: location,
             language: Self.configuration.language,
@@ -157,7 +157,7 @@ final public class WeatherService: Sendable {
     /// `let current = try await service.weather(for: newYork, including: .current)`
     ///
     final public func weather<T>(
-        for location: Location,
+        for location: LocationProtocol,
         including dataSet: WeatherQuery<T>
     ) async throws -> T {
         let proxy = try await networkClient.fetchWeather(
@@ -182,7 +182,7 @@ final public class WeatherService: Sendable {
     /// `let (current, minute) = try await service.weather(for: newYork, including: .current, .minute)`
     ///
     final public func weather<T1, T2>(
-        for location: Location,
+        for location: LocationProtocol,
         including dataSet1: WeatherQuery<T1>,
         _ dataSet2: WeatherQuery<T2>
     ) async throws -> (T1, T2) {
@@ -202,7 +202,7 @@ final public class WeatherService: Sendable {
     }
 
     final public func weather<T1, T2, T3>(
-        for location: Location,
+        for location: LocationProtocol,
         including dataSet1: WeatherQuery<T1>,
         _ dataSet2: WeatherQuery<T2>,
         _ dataSet3: WeatherQuery<T3>
@@ -225,7 +225,7 @@ final public class WeatherService: Sendable {
     }
 
     final public func weather<T1, T2, T3, T4>(
-        for location: Location,
+        for location: LocationProtocol,
         including dataSet1: WeatherQuery<T1>,
         _ dataSet2: WeatherQuery<T2>,
         _ dataSet3: WeatherQuery<T3>,
@@ -251,7 +251,7 @@ final public class WeatherService: Sendable {
     }
 
     final public func weather<T1, T2, T3, T4, T5>(
-        for location: Location,
+        for location: LocationProtocol,
         including dataSet1: WeatherQuery<T1>,
         _ dataSet2: WeatherQuery<T2>,
         _ dataSet3: WeatherQuery<T3>,
@@ -280,7 +280,7 @@ final public class WeatherService: Sendable {
     }
 
     final public func weather<T1, T2, T3, T4, T5, T6>(
-        for location: Location,
+        for location: LocationProtocol,
         including dataSet1: WeatherQuery<T1>,
         _ dataSet2: WeatherQuery<T2>,
         _ dataSet3: WeatherQuery<T3>,
